refactor: 로또 결과 객체에서 수익률, 등수 결과 객체 분리
