+++ src/main/java/lotto/result/LottoRanksResult.java
@@ -1,0 +1,19 @@
+package lotto.result;
+
+import java.util.List;
+import lotto.GameMessage;
+import lotto.domain.Rank;
+
+public class LottoRanksResult {
+    public static void showRanksStats(List<Integer> lottoRanks) {
+        for (Rank rank : Rank.values()) {
+            showRank(lottoRanks, rank);
+        }
+    }
+
+    private static void showRank(List<Integer> lottoRanks, Rank rank) {
+        if (rank != Rank.ZERO) {
+            System.out.println(rank.toString() + lottoRanks.get(rank.getRank()) + GameMessage.RESULT_RANK_UNITS);
+        }
+    }
+}
